Fix database persistent and add courses

+++ src/main/java/com/college/UserRepository.java
@@ -2,7 +2,6 @@ package com.college;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
     User findByUsername(String username);
 }
