+++ src/main/java/com/college/repository/CourseRepository.java
@@ -1,11 +1,9 @@
 package com.college.repository;
 
-import com.college.model.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.college.model.Course;
 import java.util.List;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
-
+public interface CourseRepository extends JpaRepository<Course, Long> {
     List<Course> findByDepartmentIgnoreCase(String department);
-
 }
